Remove pollercontroller unit tests
Motivation
----------
The pollercontroller unit tests are inherently racey due to the
nature of requiring parts to run in goroutines. This is leading
to test failures due to tests not behaving in the manner that
is expected. It's even clear what TestPollerControllerForceHTTPAndNewConfig
is testing anymore, if anything. What the unit tests test for
is actually covered under normal operation whilst the integration
test suite is run also.

Changes
-------
Remove pollercontroller unit tests.

Change-Id: I6c8e57391044629a2d11f41ccb2ce970503062db
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/196720
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>