GOCBC-1475: Retry cid unknown operations against default collection
Motivation
----------
When we receive a collection unknown response for an operation then
we queue the request and perform a get cid, until the operation
is cancelled/times out. This is not the case for the default
collection - a collection unknown response is immediately returned
to the user. A collection unknown can occur when a user lacks
permissions as well as when the collection does not exist. We should
update the error handling to retry the request, but not a cid refresh.

Changes
-------
Update collection unknown error handling to retry operations against
default collection.

Change-Id: Ie78acd4ff9be4bd731caf2a6a521148c2eb85576
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/196728
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>