GOCBC-1477: Update json atr attempt expiry to be int
Motivation
-----------
At the moment we have the ExpiryTime field jsonAtrAttempt as a
uint. This means that if the expiry time is somehow negative in
the atr metadata then we cannot clean up the document as parsing
the metadata json will fail.

Changes
--------
Update jsonAtrAttempt so that ExpiryTime is int rather than uint.

Change-Id: Iaa8a70e0b94f8aaa931130baf522440c062c26e3
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/196790
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>