GOCBC-1480: Update cluster map notif handling
Motivation
----------
When we get a cluster map notif showing that there is a config
newer than ours then we request the config from each node until
we are successful. We shouldn't send to nodes that don't understand
known versions. We also shouldn't consider any fetch successful
unless the SDK successfully applies the config.

Changes
-------
Update cluster map notif handling so that we only send requests
back to nodes which understand known versions.
When we get a config for a fetch originating from a cluster map
notif then only consider it successful if the config is successfully
applied.

Change-Id: I37ac41b9cb5a46196a225640783fbd5aa397eca0
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/197109
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>