GOCBC-1481: Handle config-only response status
Motivation
----------
There is an edge case where a node can be failed over but not
yet removed from the cluster, and we miss the config update
showing its removal when using cluster map notifs. When this
happens we can end up sending a kv request to a node that is in
"config-only" mode. When a node is in this mode it will reject
kv commands with a 0x0d status code. We will not recover from
this until the cluster topology undergoes another change and sends
us an updated config.
When we receive a 0x0d status we should fetch a new config from
the server and retry the request.

Changes
-------
Add status code for 0x0d to memd.
Add corresponding errors for the status code.
When we receive a config-only error then fetch a new config
with our current revision numbers, and retry the request after
the config has been applied.

Change-Id: I47477688218731ab692d27d8c738ab2dd7f1579d
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/197186
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>