GOCBC-1483: Retry requests when resetting cid cache queue
Motivation
-----------
When kv operations are written to collection that
we do not know then we will queue the operations
and refresh the collection id. If that cid refresh
fails with a collection unknown error then we will
repeatedly retry fetching until all operations in
the queue timeout or we receive a cid. In the case
that there are multiple operations already queued
when we receive the collection unknown then we
will drop all of the operations other than the
initial operation. These operations will then sit
until they are timed out/cancelled.

Changes
--------
On unknownCid detection drain the current opQueue into
the new opQueue.

Change-Id: I857b8e25d3c1d888dc64ef94b03548fcc134dfa1
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/197437
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>