GOCBC-1484: Start cmd span on range scan cancel
Motivation
----------
When we create a range scan cancel request we
bypass the usual dispatch pipeline, which creates
a CMD span on the request. We need to create the
CMD span as a part of cancel, the same as continue.

Changes
-------
Start a cmd span on the request as a part of range
scan cancel.

Change-Id: I103e366ed583137ee1c9b26bcf54543c1a5d1c59
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/212816
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>