GOCBC-1484: Start cmd span on range scan cancel
Motivation
----------
When we create a range scan cancel request we
bypass the usual dispatch pipeline, which creates
a CMD span on the request. We need to create the
CMD span as a part of cancel, the same as continue.

Changes
-------
Start a cmd span on the request as a part of range
scan cancel.

Change-Id: Ib12675a1729cb1ff3bfe2f1c7f3cc1ac0d97b18c
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/197551
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>