GOCBC-1542: Drop "identical revision" log line to debug level
Motivation
-----------
When we receive a cluster config that is identical to one that
we already know then we log that information. We recently raised
the log level for this line to info but unfortunately it is
creating too much noise when a large number of agents are in use,
we need to drop this back to debug level.

Changes
--------
Drop "identical revision" log line to debug level.

Change-Id: I9bb543385486d82b272107203d793ebee402fd84
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/201281
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>