GOCBC-1542: Drop "identical revision" log line to debug level
Motivation
-----------
When we receive a cluster config that is identical to one that
we already know then we log that information. We recently raised
the log level for this line to info but unfortunately it is
creating too much noise when a large number of agents are in use,
we need to drop this back to debug level.

Changes
--------
Drop "identical revision" log line to debug level.

Change-Id: I5824e87be8c84e184cb9b229c08f3033f0036ae5
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/201278
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>