GOCBC-1550: Release lock when buildFirstRouteConfig returns nil
Motivation
----------
If the first time that we build a routeconfig returns nil then
we do not release the config lock.  This can only happen if the
SDK is in ns_server mode and the server provides us with a config
which does not contain a thisNode entry.

Changes
-------
Update onNewConfig to release the lock if buildFirstRouteConfig
returns nil.
Update buildFirstRouteConfig to return an invalid route config
to match buildRouteConfig.

Change-Id: I3e7adb7e4c5ccd92602ecd50387689c154714a3c
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/201617
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/201684
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>