GOCBC-1561: Use error map description for unknown kv statuses
Motivation
-----------
When we receive a kv status code that we do not know we end
up returning an error to the user with the usual error
context but a message like "unknown kv status code (0x01)".
We should improve this by taking the error description from
the error map and using that as the message when we don't
recognise the status code.

Changes
-------
When we receive an unknown status code from kv consult the
error map, if the error map contains the error then use
the description as the error message.

Change-Id: I8ad4e0feacbef7489663ab7caa37241897fe9b86
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/202540
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>