GOCBC-1569: Reload pipeline if dispatch direct fails with closed
Motivation
----------
In DispatchDirectToAddress we retry the dispatch if it fails
with a pipeline closed error. In the case that this error is
returned then retrying will always return the same error as it
uses the same pipeline.

Changes
--------
Update DispatchDirectToAddress to accept an address rather than a
pipeline and perform a lookup in the mux state each dispatch
attempt.

Change-Id: I1c5920e84fc48a8e718bb5cdb41c3ce006a36a1e
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/203138
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>