GOCBC-1569: Reload pipeline if dispatch direct fails with closed
Motivation
----------
In DispatchDirectToAddress we retry the dispatch if it fails
with a pipeline closed error. In the case that this error is
returned then retrying will always return the same error as it
uses the same pipeline.

Changes
--------
Update DispatchDirectToAddress to accept an address rather than a
pipeline and perform a lookup in the mux state each dispatch
attempt.

Change-Id: Id6be1d02ebaa850f91ad31c1f15249b31647f87b
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/203474
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>