GOCBC-1622: Update degraded handling for wait until ready
Motivation
----------
WaitUntilReady in degraded state currently defaults to waiting for
all services to be ready on at least one node. This means that
degraded can never succeed if one of those services does not exist
on the cluster.

Changes
-------
Update degraded status to only wait for available services by
default.

Change-Id: I757d618b9cf41306c39e2ffa861c432b9f628ca0
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/207678
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>