GOCBC-1596: Wait for any in-progress lost txn cleaner processes before removing client
Motivation
==========
Currently it is possible to have a race between removeClient and client records being updated/created which means that the client may be added again to the client record. Even though this is acceptable, as the spec considers removeClient to be non-essential (the client will always get removed by another client once it expires), this can be problematic for Cleanup testing in FIT where it is convenient in each test case to assume that it starts with a clean slate. Waiting for any in-progress creates/updates before removing the client should reduce the chance of such a race happening.

Change
======
Wait for any in-progress process calls to finish before removing the client from the record.

Change-Id: Ia9035d538625601784194ea22005b217499e283f
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/207704
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>