GOCBC-1631: Handle scope not found status code
Motivation
-----------
We have code for converting the 0x8c memcached status to a
scope not status, and we have scope not found errors. However,
we do not have code linking the two together. We should add this
glue, and also handle scope not found internally in the same places
that we handle collection not found.

Changes
-------
Wire up translation from scope not found status code to error.
Handle scope not found for retries, the same as collection not
found.

Change-Id: Ie265c529d841d480068269f6a6c3e294f64656be
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/210416
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>