GOCBC-1626: Update behaviour of ping and http reqs on agent close
Motivation
----------
After agent close we currently allow http requests to continue to
be sent. Ping will also loop with an error on the kv side until
timeout.

Changes
-------
Update http component to cancel in flight requests when close
is called.
Update http component so that calls after close will return an
error.
Update ping and diagnostics to populate results with errors if the
agent is already closed.

Change-Id: I4d3e0d1c49e1f2e9f31cdc3a61409decb733e033
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/210686
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>