GOCBC-1636: Copy clusteragent config watchers on new config
Motivation
-----------
There is a data race between remove/add config watcher and
OnNewRouteConfig in the cluster agent. This happens because we
aren't copying the config watchers before using them, rather we
just assign a variable to point at the same locaiton.

Changes
-------
Update clusteragent OnNewRouteConfig so that it copies the config
watchers before using them.

Change-Id: I5b89e1a23fadb623bcb2460b1c4a5762bdad4a32
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/210791
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>