GOCBC-1632: Add support for server groups
Motivation
----------
Operations do not currently take into account server groups. For
reading from replica servers we need to provide a way for users to
be able to read from specific groups, which can correspond to
availablity zones. Doing so helps users to reduce costs of sending
requests across availability zones.

Changes
-------
Update config snapshot to provide a way to get all servers for a
given key in a given server group.
Add a server group option to getOneReplica and LookupIn.
Add logic in request routing to reject requests that are to
nodes that aren't in a specified server group.

Change-Id: I421a73184910ba6fd8684011fea90f21413dc62c
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/211522
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>