GOCBC-1645: Reduce output from CCCP poller when configs are empty
Motivation
----------
When we added support for cluster map known versions we added logic
to the cccp poller to log that we'd received an empty block and
then if all nodes returned the same we'd log again. This is quite
verbose and unneccesary.

Changes
-------
Update cccp poller to remove log line about empty block.

Change-Id: Ic6a41ba6c73bc15dee5c4382cfa8baa54034e806
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/212559
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>