GOCBC-1659: Do not attempt to refresh config in ns_server mode when receiving config-only KV status
Motivation
==========
When receiving a 0x0d (config-only) KV status, we attempt to refresh the config using the cccpConfigFetcher. In ns server mode, the configFetcher is not set, resulting in a nil pointer dereference.

Change
======
Do not refresh the config when handling a config-only KV status in ns server mode (i.e. in the configManagementComponent, check whether the configFetcher is nil before using it).

Change-Id: Ic0ea6182e7bcb0479151cb00479bd45344b82c75
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/215277
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>