GOCBC-1660: Fix debug log data race when closing KV mux
Motivation
==========
When closing the KV mux we log all fields of the memdPipeline structs. This includes a lock that is acquired when fetching a new config, which results in a data race. The data race was detected via TestCollectionHistoryRetentionUnsupported.

Change
======
Selectively include certain fields from memdPipeline instead of logging everything (to avoid logging the lock's value)

Change-Id: I9aeb18f033b27d3b6c946e6609c56d48ae748cfa
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/215805
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Build Bot <build@couchbase.com>