GOCBC-1311: Add support for transactions ExtThreadSafety
Motivation
-----------
Gocbcore was the basis for the ExtThreadSafety spec, however
there are a few updates that we need to make to bring us in line
with the spec.

Changes
-------
Add support for operation id to the transaction metadata, which is
used in stored mutations and written as a part of inserts.
Add two new error types to replace previous operation failed in
some places.

Change-Id: I1bff8421de3fd22b2dfd318c28c0fe17b2272316
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/181212
Tested-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>