GOCBC-1677: Include vbucket id in trace logs when writing request
Motivation
----------
There are cases when debugging where knowing the vbucket id is
very useful. Especially in cases where we get a not my vbucket
error but the key is redacted.

Changes
-------
Include vbucket id in trace logs when writing request.
Add a new trace level log when we receive a not my vbucket error.

Change-Id: Id5677589ff7279c416c2b85df9e6da53f67f84b4
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/219573
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>