GOCBC-1680: Improve retry handling for http requests
Motivation
-----------
There are times when we try to send a http request and receive an
error indicative of something like a failed dial. In the case of
errors like this we should attempt a retry.

Changes
-------
Add a check for net.OpError and if the op is a dial then assign
a SocketNotAvailableRetryReason else if the op is anything else
then assign SocketCloseInFlightRetryReason.
Add a check for net.DnsError and assign SocketNotAvailableRetryReason.
Update maybeWait to return a timeout rather than the underlying error
if the deadline is exceeded.

Change-Id: Ie80740b1014fa18f4ba2746abd7c20d0a6aad98e
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/219807
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>