GOCBC-1711: Ensure that the websocket telemetry connection is closed when closing the TelemetryReporter
Motivation
==========
When receiving the shutdownSignal we were only cancelling the context that was used to dial the websocket connection. That doesn't cause the connection to be closed after it has been successfsully established.

Changes
=======
* Make the context only be within the scope of dialling the connection
* Add a separate listener for shutdownSignal that will close the connection once the channel is closed.
* Clean up the error handling to remove unnecessary references to context.Canceled.

Results
=======
The websocket connection is now being closed. Verified by the absence of a goroutine leak in gocb tests & from logs.

Change-Id: I3fdcb54739cf5c13272e4b6ac485d296d26b87ce
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/225117
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Build Bot <build@couchbase.com>