GOCBC-1717: Update rollback failures to set state bits
Motivation
----------
As a part of ExtThreadSafety we updated rollback so that failures
do not set state bits. This wasn't quite accurate to the spec, which
expects behaviour at the core loop level to handle this behaviour
change. This isn't compatible with direct users of gocbcore so
we should continue to set the state bits, which is still in keeping
with the behaviour in the spec.

Changes
--------
Update rollback failures to set state bits.

Change-Id: I645b2c360d50213d9406c1d7761a33ba564f663b
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/226004
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>