GOCBC-1733: Save CAS to stagedMutation after call to AfterDocCommittedBeforeSavingCAS txn hook
Motivation
==========
We are not saving the CAS value to the stagedMutation between calls to AfterDocCommittedBeforeSavingCAS and AfterDocCommitted txn hooks. The names of the hooks imply that we should be doing this, otherwise calls to both hooks would have the same outcome.

This shouldn’t affect behaviour from the users' perspective. It only matters when the testing hooks are in use.

Change
======
Save CAS to stagedMutation after call to AfterDocCommittedBeforeSavingCAS txn hook

Change-Id: I8e890721c40a53b2bfbe3982595bbda782cf6c61
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/228330
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>