GOCBC-1742: Do not wrap ErrDocumentExists in TOF when doc exists as staged mutation
Motivation
==========
Since ExtInsertExisting, the spec expects us to return unwrapped ErrDocumentExists errors & not set internal state, allowing users to handle the error and prevent the transaction from failing.

Change
======
In cases where the document exists as a staged mutation, return an unwrapped ErrDocumentExists, similar to the error we return when the document is found on the server.

Results
=======
All relevant FIT tests now pass.

Change-Id: I62c85f45fd42ea322b017a522d2a47846afc382f
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/229260
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>