GOCBC-1747: Change how srv refresh is triggered
Motivation
-----------
We've utilised CCCP being unable to fetch a config from any
node as a trigger to refresh the srv record, if in use. When we
connect to servers that support config push we no longer CCCP poll
which means that we no SRV refresh.

Changes
-------
Add events for clients connecting and disconnecting and keep
counts of number of clients and pipelines connected.
When all pipelines disconnect trigger the SRV refresh logic.

Change-Id: Ic4a70a5c9968c16b0a732f5ad181987a85424d28
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/229883
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>