GOCBC-1693: Remove telemetryOpRecorder and use telemetryComponent directly for KV ops
Motivation
==========
Previously we were creating a telemetryOpRecorder for each memd request, each wrapping a pointer to the TelemetryReporter. As we want to also report metrics using request tombstones, tombstones would have to include this pointer to the TelemetryReporter.
We should store the attributes to the memdQRequest and the memdClient that is being used, and use the telemetryComponent via the memdClient, to avoid the tombstones carrying pointers to the TelemetryReporter.

Changes
=======
* Pass down the telemetryComponent, nodeUUID and canonicalAddress to the memdClient.
* Store last dispatch time in memdQRequestConnectionInfo before writing the request to the connection, which can be used to calculate the response latency.
* Report the telemetry metrics entry via the telemetryComponent held by the memdClient.

Results
=======
All tests pass

Change-Id: I8543897ecf60bd340a73e9f8f3410785dbc22ac1
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/230125
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>