GOCBC-1755: Adjust analytics prodName check to prod
Motivation
----------
The server has added new prod field to the cluster config, which
we should use instead of prodName. We should also check for the
value "analytics".

Changes
-------
Update config to contain a prod field instead of prodName.
Update the enterprise analytics block check to check for
a value of "analytics".
Update agentgroup to pass InternalConfig to cluster agent.

Change-Id: I40d988efbb75a1370b1ae8009ed2450cdecfd4c0
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/231142
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>